% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Compute_ModisLST.R
\name{Compute_ModisLST}
\alias{Compute_ModisLST}
\title{Compute MODIS Land Surface Temperature}
\usage{
Compute_ModisLST(path_to_tif, path_to_shapefiles, path_mod_shapefile,
  aggregate)
}
\arguments{
\item{path_to_tif}{Character String; Path to the folder that contains the tif files}

\item{path_to_shapefiles}{;Character String or SpatialPolygonsDataFrames object; The path to the shapefile (.shp) or the shapefile object}

\item{path_mod_shapefile}{;Character String ; The path where the user wants the Modified Shapefile to be stored}

\item{aggregate}{;Character String ; Aggregate values  1. daily
2. weekly
3. monthly
4. yearly}
}
\value{
Dataframe with new columns appended with Land Surface Temperature in Celcius and modified shapefiles.
}
\description{
Function automates the computation of MODIS Land Surface Temperature.It creates Shapefiles with Temperature columns appended and it is stored at the path provided by the user.This requires the user to download the raster files first.
The user must use Download_tif to download the required raster files.
}
\examples{
shape_dsn <- system.file("vectors", package = "LSTModis")[1]
tif_dsn <- system.file("pictures", package = "LSTModis")[1]
Shapefile <- rgdal::readOGR(dsn=shape_dsn, layer="Shapefile")
pwd <- getwd()
df<-Compute_ModisLST(path_to_tif = tif_dsn ,
          path_to_shapefiles = Shapefile ,
          path_mod_shapefile = pwd,
          aggregate = "weekly")
}
